+++ app/src/main/java/com/ninebx/ui/base/realm/Users.java
@@ -51,13 +51,13 @@ public class Users extends RealmObject implements Parcelable {
     @Required private String country              = "";
     @Required private String userId               = "";
     @PrimaryKey //@Required
-    private int id                  = 0;
+    private long id                  = 0;
     @Required private Boolean isCompleteProfile   = false;
     @Required private String profilePhoto         = "";
     @Required private RealmList<Member> members = new RealmList<Member>();
     @Ignore private RealmList<Member> decryptedMembers = new RealmList<Member>();
 
-    public Users(String fullName, String emailAddress, String relationship, String dateOfBirth, String anniversary, String gender, String mobileNumber, String street_1, String street_2, String city, String state, String zipCode, String country, int id, RealmList<Member> members) {
+    public Users(String fullName, String emailAddress, String relationship, String dateOfBirth, String anniversary, String gender, String mobileNumber, String street_1, String street_2, String city, String state, String zipCode, String country, long id, RealmList<Member> members) {
         this.fullName = fullName;
         this.emailAddress = emailAddress;
         this.relationship = relationship;
@@ -79,7 +79,7 @@ public class Users extends RealmObject implements Parcelable {
     }
 
 
-    public Users(String fullName, String emailAddress, String relationship, String dateOfBirth, String anniversary, String gender, String mobileNumber, String street_1, String street_2, String city, String state, String zipCode, String country, String userId, int id, RealmList<Member> members) {
+    public Users(String fullName, String emailAddress, String relationship, String dateOfBirth, String anniversary, String gender, String mobileNumber, String street_1, String street_2, String city, String state, String zipCode, String country, String userId, long id, RealmList<Member> members) {
         this.fullName = fullName;
         this.emailAddress = emailAddress;
         this.relationship = relationship;
@@ -282,11 +282,11 @@ public class Users extends RealmObject implements Parcelable {
         this.country = country;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
