+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedClothingSizes.java
@@ -31,7 +31,7 @@ public class DecryptedClothingSizes implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-    private int id = 0;
+    private long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -268,7 +268,7 @@ public class DecryptedClothingSizes implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(classType);
@@ -329,11 +329,11 @@ public class DecryptedClothingSizes implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
