+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedEmergencyContacts.java
@@ -31,7 +31,7 @@ public class DecryptedEmergencyContacts implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-            int id = 0;
+            long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -156,7 +156,7 @@ public class DecryptedEmergencyContacts implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(classType);
@@ -184,11 +184,11 @@ public class DecryptedEmergencyContacts implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
