+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedHealthcareProviders.java
@@ -31,7 +31,7 @@ public class DecryptedHealthcareProviders implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-            int id = 0;
+            long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -132,7 +132,7 @@ public class DecryptedHealthcareProviders implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(classType);
@@ -161,11 +161,11 @@ public class DecryptedHealthcareProviders implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
