+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedIdentification.java
@@ -31,7 +31,7 @@ public class DecryptedIdentification implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-            int id = 0;
+            long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -128,7 +128,7 @@ public class DecryptedIdentification implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(classType);
@@ -156,11 +156,11 @@ public class DecryptedIdentification implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
