+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedInterests.java
@@ -30,7 +30,7 @@ public class DecryptedInterests implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-    private int id = 0;
+    private long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -139,7 +139,7 @@ public class DecryptedInterests implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(institutionName);
@@ -170,11 +170,11 @@ public class DecryptedInterests implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
