+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedLoyaltyPrograms.java
@@ -31,7 +31,7 @@ public class DecryptedLoyaltyPrograms implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-    private int id = 0;
+    private long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -112,7 +112,7 @@ public class DecryptedLoyaltyPrograms implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(brandName);
@@ -136,11 +136,11 @@ public class DecryptedLoyaltyPrograms implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
