+++ app/src/main/java/com/ninebx/ui/base/realm/decrypted/DecryptedMedicalConditions.java
@@ -31,7 +31,7 @@ public class DecryptedMedicalConditions implements Parcelable {
         }
     };
     @PrimaryKey //@Required
-            int id = 0;
+            long id = 0;
     @Required
     private RealmList<RealmString> backingImages = new RealmList<>();
     @Ignore
@@ -80,7 +80,7 @@ public class DecryptedMedicalConditions implements Parcelable {
     public DecryptedMedicalConditions() {
     }
 
-    public DecryptedMedicalConditions(int id, RealmList<RealmString> backingImages, List<String> photosId, String selectionType, String classType, String condition, String dateDiagnosed, String medi_description, String notes, String attachmentNames, String created, String modified, Boolean isPrivate, String createdUser) {
+    public DecryptedMedicalConditions(long id, RealmList<RealmString> backingImages, List<String> photosId, String selectionType, String classType, String condition, String dateDiagnosed, String medi_description, String notes, String attachmentNames, String created, String modified, Boolean isPrivate, String createdUser) {
         this.id = id;
         this.backingImages = backingImages;
         this.photosId = photosId;
@@ -99,7 +99,7 @@ public class DecryptedMedicalConditions implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeStringList(photosId);
         dest.writeString(selectionType);
         dest.writeString(classType);
@@ -119,11 +119,11 @@ public class DecryptedMedicalConditions implements Parcelable {
         return 0;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId( long id ) {
         this.id = id;
     }
